run remixd only for remixd tests
