+++ apps/remix-ide/ci/browser_test.sh
@@ -11,7 +11,6 @@ TEST_EXITCODE=0
 yarn run ganache-cli &
 yarn run serve:production &
 echo 'sharing folder: ' $PWD '/apps/remix-ide/contracts' &
-yarn run remixd -s ./apps/remix-ide/contracts &
 
 sleep 5
